moving useTip to api module
